tests/wpt/tests/websockets/interfaces/WebSocket/constants/003.html

<!doctype html>
<title>WebSockets: deleting constants</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=../../../constants.sub.js></script>
<meta name="variant" content="">
<meta name="variant" content="?wss">
<div id=log></div>
<script>
var constants = ['CONNECTING', 'OPEN', 'CLOSING', 'CLOSED'];
for (var i = 0; i < constants.length; ++i) {
  test(function(){
    var ws = new WebSocket(SCHEME_DOMAIN_PORT+'/');
    delete WebSocket[constants[i]]; // should be ignored, has { DontDelete }
    delete WebSocket.prototype[constants[i]]; // should be ignored, has { DontDelete }
    delete ws[constants[i]]; // should be ignored, there is no such property on the object
    assert_equals(WebSocket[constants[i]], i, 'WebSocket.'+constants[i]);
    assert_equals(WebSocket.prototype[constants[i]], i, 'WebSocket.prototype.'+constants[i]);
    assert_equals(ws[constants[i]], i, 'ws.'+constants[i]);
  })
};
</script>
